agent/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-cartesia
python-dotenv
requests
Fastapi 
"pymongo[srv]"==3.12
bcrypt 
python-dotenv 
uvicorn
